requirements.txt
Pillow==11.2.1
streamlit==1.45.1
torch==2.7.1
transformers==4.52.4
